use own RMap
