+++ storage/Cargo.toml
@@ -24,7 +24,6 @@ prometheus-client = { workspace = true }
 futures-util = { workspace = true }
 tracing = { workspace = true }
 zstd = { workspace = true }
-rangemap = "1.5.1"
 
 [dev-dependencies]
 tracing-subscriber = { workspace = true }
