+++ storage/src/lib.rs
@@ -12,3 +12,4 @@ pub mod index;
 pub mod journal;
 pub mod metadata;
 pub mod mmr;
+pub mod rmap;
